+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Changelog
  * removed `InstallInfo::*Enabled*()` methods
  * added config key "bootstrap-file" which is loaded before loading plugins
  * added repository type "path-mapping"
+ * added config key "repository.optimize"
 
 * 1.0.0-beta5 (2015-05-29)